fix: auto deregister hotkeys when focused app is excluded
We have double check for excluded app. If an app is excluded, when we
focus on that app, it will try to deregister all the hotkeys, so that we
can use the hotkey for something else, and auto register back when focus
on non-excluded app. If that didn't work, mode activation also checks
for it as a second guard.